Create anonymous class override to reduce cyclomatic complexity
