+++ packages/framework/tests/Feature/Commands/RouteListCommandTest.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature\Commands;
 
+use Hyde\Hyde;
+use Hyde\Pages\InMemoryPage;
+use Hyde\Support\Models\Route;
 use Hyde\Testing\TestCase;
 
 /**
@@ -30,5 +33,47 @@ class RouteListCommandTest extends TestCase
                     'index',
                 ],
             ])->assertExitCode(0);
+    }
+
+    public function testConsoleRouteList()
+    {
+        Hyde::routes()->forget('404');
+
+        $this->artisan('route:list')
+            ->expectsTable(['Page Type', 'Source File', 'Output File', 'Route Key'], [[
+                'page_type' => 'BladePage',
+                'source_file' => '<href=file://'.str_replace('\\', '/', Hyde::path()).'/_pages/index.blade.php>_pages/index.blade.php</>',
+                'output_file' => '_site/index.html',
+                'route_key' => 'index',
+            ]])->assertExitCode(0);
+    }
+
+    public function testConsoleRouteListWithClickableOutputPaths()
+    {
+        Hyde::routes()->forget('404');
+        $this->file('_site/index.html');
+
+        $this->artisan('route:list')
+            ->expectsTable(['Page Type', 'Source File', 'Output File', 'Route Key'], [[
+                'page_type' => 'BladePage',
+                'source_file' => '<href=file://'.str_replace('\\', '/', Hyde::path()).'/_pages/index.blade.php>_pages/index.blade.php</>',
+                'output_file' => '<href=file://'.str_replace('\\', '/', Hyde::path()).'/_site/index.html>_site/index.html</>',
+                'route_key' => 'index',
+            ]])->assertExitCode(0);
+    }
+
+    public function testConsoleRouteListWithDynamicPages()
+    {
+        Hyde::routes()->forget('404');
+        Hyde::routes()->forget('index');
+        Hyde::routes()->put('foo', new Route(new InMemoryPage('foo')));
+
+        $this->artisan('route:list')
+            ->expectsTable(['Page Type', 'Source File', 'Output File', 'Route Key'], [[
+                'page_type' => 'InMemoryPage',
+                'source_file' => '<fg=yellow>dynamic</>',
+                'output_file' => '_site/foo.html',
+                'route_key' => 'foo',
+        ]])->assertExitCode(0);
     }
 }
